stupidly_simple_ocr_app_for_practice/pyproject.toml
[tool.poetry]
name = "ocr_app"
version = "0.1.0"
description = ""
authors = ["Sergey Balashov <balashevski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pytesseract = "^0.3.10"
pillow = "^10.3.0"
requests = "^2.32.3"
numpy = "^2.0.0"
opencv-python-headless = "^4.10.0.82"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
